feat: :art: Add tests and make the things work
I wanted a way to quickly iterate on this and so added a local version of the transaction payload as well as nodemon to continually run the action on file save. This now seems to work with the bon event and gets back enough info to send to chain vine.